Controller file structure updated
